feat: add fn that checks correct route
