+++ src/shared/services/validate-product-type.ts
@@ -1,0 +1,8 @@
+import { ALLOWED_PRODUCT_TYPE, type ProductType } from "../api/routes";
+
+export function validateProductType(productType: string | undefined): ProductType | null {
+  if (productType && ALLOWED_PRODUCT_TYPE.includes(productType as ProductType)) {
+    return productType as ProductType;
+  }
+  return null;
+}
